requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
plotly==5.17.0
pandas==2.1.4

